Move core CLI formatting infrastructure to `workspace`
