+++ crates/air/Cargo.toml
@@ -12,20 +12,13 @@ repository.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-air_r_formatter = { workspace = true }
-air_r_parser = { workspace = true }
 anyhow = { workspace = true }
-biome_formatter = { workspace = true }
-biome_parser = { workspace = true }
 clap = { workspace = true, features = ["wrap_help"] }
 colored = { workspace = true }
 crates = { workspace = true }
 fs = { workspace = true }
-ignore = { workspace = true }
 itertools = { workspace = true }
-line_ending = { workspace = true }
 lsp = { workspace = true }
-thiserror = { workspace = true }
 tokio = "1.41.1"
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
