+++ crates/air/src/commands/format.rs
@@ -1,23 +1,13 @@
-use std::fmt::Display;
-use std::fmt::Formatter;
 use std::io;
 use std::io::stderr;
 use std::io::Write;
-use std::path::Path;
-use std::path::PathBuf;
 
-use air_r_formatter::context::RFormatOptions;
-use air_r_parser::RParserOptions;
-use colored::Colorize;
-use fs::relativize_path;
-use itertools::Either;
 use itertools::Itertools;
-use thiserror::Error;
-use workspace::discovery::discover_r_file_paths;
 use workspace::discovery::discover_settings;
 use workspace::discovery::DiscoveredSettings;
+use workspace::format::FormattedFile;
+use workspace::format::OnChangedAction;
 use workspace::resolve::PathResolver;
-use workspace::settings::FormatSettings;
 use workspace::settings::Settings;
 
 use crate::args::FormatCommand;
@@ -27,8 +17,8 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
     let mode = FormatMode::from_command(&command);
 
     let on_changed_action = match mode {
-        FormatMode::Write => FormatFileOnChangedAction::Write,
-        FormatMode::Check => FormatFileOnChangedAction::None,
+        FormatMode::Write => OnChangedAction::Write,
+        FormatMode::Check => OnChangedAction::None,
     };
 
     let mut resolver = PathResolver::new(Settings::default());
@@ -41,7 +31,8 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
         resolver.add(&directory, settings);
     }
 
-    let (actions, errors) = format_paths(&command.paths, &resolver, on_changed_action);
+    let (files, errors) =
+        workspace::format::format_paths(&command.paths, &resolver, on_changed_action);
 
     for error in &errors {
         tracing::error!("{error}");
@@ -50,7 +41,7 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
     match mode {
         FormatMode::Write => {}
         FormatMode::Check => {
-            write_changed(&actions, &mut stderr().lock())?;
+            write_changed(&files, &mut stderr().lock())?;
         }
     }
 
@@ -64,7 +55,7 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
         }
         FormatMode::Check => {
             if errors.is_empty() {
-                let any_changed = actions.iter().any(FormatFileAction::is_changed);
+                let any_changed = files.iter().any(FormattedFile::is_changed);
 
                 if any_changed {
                     Ok(ExitStatus::Failure)
@@ -76,28 +67,6 @@ pub(crate) fn format(command: FormatCommand) -> anyhow::Result<ExitStatus> {
             }
         }
     }
-}
-
-fn format_paths<P: AsRef<Path>>(
-    paths: &[P],
-    resolver: &PathResolver<Settings>,
-    on_changed_action: FormatFileOnChangedAction,
-) -> (Vec<FormatFileAction>, Vec<FormatFileError>) {
-    let paths = discover_r_file_paths(paths, resolver, true);
-
-    paths
-        .into_iter()
-        .map(|path| match path {
-            Ok(path) => {
-                let settings = resolver.resolve_or_fallback(&path);
-                format_file(path, on_changed_action, &settings.format)
-            }
-            Err(err) => Err(err.into()),
-        })
-        .partition_map(|result| match result {
-            Ok(result) => Either::Left(result),
-            Err(err) => Either::Right(err),
-        })
 }
 
 #[derive(Copy, Clone, Debug)]
@@ -116,12 +85,12 @@ impl FormatMode {
     }
 }
 
-fn write_changed(actions: &[FormatFileAction], f: &mut impl Write) -> io::Result<()> {
-    for path in actions
+fn write_changed(files: &[FormattedFile], f: &mut impl Write) -> io::Result<()> {
+    for path in files
         .iter()
         .filter_map(|result| match result {
-            FormatFileAction::Changed(path) => Some(path),
-            FormatFileAction::Unchanged => None,
+            FormattedFile::Changed(path) => Some(path),
+            FormattedFile::Unchanged => None,
         })
         .sorted_unstable()
     {
@@ -129,158 +98,6 @@ fn write_changed(actions: &[FormatFileAction], f: &mut impl Write) -> io::Result
     }
 
     Ok(())
-}
-
-#[derive(Debug, Clone, Copy)]
-pub(crate) enum FormatFileOnChangedAction {
-    Write,
-    None,
-}
-
-pub(crate) enum FormatFileAction {
-    Changed(PathBuf),
-    Unchanged,
-}
-
-impl FormatFileAction {
-    fn is_changed(&self) -> bool {
-        matches!(self, FormatFileAction::Changed(_))
-    }
-}
-
-fn format_file(
-    path: PathBuf,
-    on_changed_action: FormatFileOnChangedAction,
-    settings: &FormatSettings,
-) -> Result<FormatFileAction, FormatFileError> {
-    tracing::trace!("Formatting {path}", path = path.display());
-
-    let source = match std::fs::read_to_string(&path) {
-        Ok(source) => source,
-        Err(err) => {
-            return Err(FormatFileError::Read(path, err));
-        }
-    };
-
-    let options = settings.to_format_options(&source);
-
-    let formatted = match format_source(source.as_str(), options) {
-        Ok(formatted) => formatted,
-        Err(err) => return Err(FormatFileError::Format(path, err)),
-    };
-
-    match formatted {
-        FormatSourceAction::Changed(new) => {
-            match on_changed_action {
-                FormatFileOnChangedAction::Write => match std::fs::write(&path, new) {
-                    Ok(()) => (),
-                    Err(err) => {
-                        return Err(FormatFileError::Write(path, err));
-                    }
-                },
-                FormatFileOnChangedAction::None => {}
-            }
-            Ok(FormatFileAction::Changed(path))
-        }
-        FormatSourceAction::Unchanged => Ok(FormatFileAction::Unchanged),
-    }
-}
-
-#[derive(Debug)]
-pub(crate) enum FormatSourceAction {
-    /// The source was formatted, and the [`String`] contains the transformed source code.
-    Changed(String),
-    /// The source was unchanged.
-    Unchanged,
-}
-
-#[derive(Error, Debug)]
-pub(crate) enum FormatSourceError {
-    #[error(transparent)]
-    Parse(#[from] air_r_parser::ParseError),
-    #[error(transparent)]
-    Format(#[from] biome_formatter::FormatError),
-    #[error(transparent)]
-    Print(#[from] biome_formatter::PrintError),
-}
-
-/// Formats a vector of `source` code
-pub(crate) fn format_source(
-    source: &str,
-    options: RFormatOptions,
-) -> std::result::Result<FormatSourceAction, FormatSourceError> {
-    let parsed = air_r_parser::parse(source, RParserOptions::default());
-
-    if parsed.has_errors() {
-        let error = parsed.into_errors().into_iter().next().unwrap();
-        return Err(error.into());
-    }
-
-    let formatted = air_r_formatter::format_node(options, &parsed.syntax())?;
-    let formatted = formatted.print()?;
-    let formatted = formatted.into_code();
-
-    if source.len() == formatted.len() && source == formatted.as_str() {
-        Ok(FormatSourceAction::Unchanged)
-    } else {
-        Ok(FormatSourceAction::Changed(formatted))
-    }
-}
-
-#[derive(Error, Debug)]
-pub(crate) enum FormatFileError {
-    Ignore(#[from] ignore::Error),
-    Format(PathBuf, FormatSourceError),
-    Read(PathBuf, io::Error),
-    Write(PathBuf, io::Error),
-}
-
-impl Display for FormatFileError {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Ignore(err) => {
-                if let ignore::Error::WithPath { path, .. } = err {
-                    write!(
-                        f,
-                        "Failed to format {path}: {err}",
-                        path = relativize_path(path).underline(),
-                        err = err
-                            .io_error()
-                            .map_or_else(|| err.to_string(), std::string::ToString::to_string)
-                    )
-                } else {
-                    write!(
-                        f,
-                        "Encountered error: {err}",
-                        err = err
-                            .io_error()
-                            .map_or_else(|| err.to_string(), std::string::ToString::to_string)
-                    )
-                }
-            }
-            Self::Read(path, err) => {
-                write!(
-                    f,
-                    "Failed to read {path}: {err}",
-                    path = relativize_path(path).underline(),
-                )
-            }
-            Self::Write(path, err) => {
-                write!(
-                    f,
-                    "Failed to write {path}: {err}",
-                    path = relativize_path(path).underline(),
-                )
-            }
-            Self::Format(path, err) => {
-                write!(
-                    f,
-                    "Failed to format {path}: {err}",
-                    path = relativize_path(path).underline(),
-                )
-            }
-        }
-    }
 }
 
 #[cfg(test)]
