+++ crates/workspace/Cargo.toml
@@ -17,10 +17,13 @@ schemars = ["dep:schemars", "settings/schemars"]
 
 [dependencies]
 air_r_formatter = { workspace = true }
+air_r_parser = { workspace = true }
 anyhow = { workspace = true }
 biome_formatter = { workspace = true, features = ["serde"] }
+colored = { workspace = true }
 fs = { workspace = true }
 ignore = { workspace = true }
+itertools = { workspace = true }
 line_ending = { workspace = true }
 rustc-hash = { workspace = true }
 thiserror = { workspace = true }
