+++ crates/workspace/src/lib.rs
@@ -7,6 +7,7 @@
 
 pub mod discovery;
 pub mod file_patterns;
+pub mod format;
 pub mod resolve;
 pub mod settings;
 pub mod toml;
